+++ packages/eas-cli/src/commands/update/view.ts
@@ -39,7 +39,8 @@ export async function viewUpdateAsync({
         `,
         {
           groupId,
-        }
+        },
+        { additionalTypenames: ['Update'] }
       )
       .toPromise()
   );