+++ packages/eas-cli/src/graphql/queries/BuildQuery.ts
@@ -54,6 +54,7 @@ export const BuildQuery = {
           { buildId },
           {
             requestPolicy: useCache ? 'cache-first' : 'network-only',
+            additionalTypenames: ['Build'],
           }
         )
         .toPromise()